+++ 09.23. 02 sikidomok/IPlaneshapes.cs
@@ -2,6 +2,6 @@
 {
     public interface IPlaneshapes //Felület létrehozva
     {
-        int  Area();
+        double Area();
     }
 }
