+++ 09.23. 02 sikidomok/Triangle.cs
@@ -13,7 +13,7 @@ namespace _09._23._02_sikidomok
             this.sideLength = sideLength;
         }
 
-        public int Area()
+        public double Area()
         {
             return (height * sideLength) / 2;
         }
